refactor: extract implicit conversion helper functions of vector
