+++ src/common/function/src/scalars/vector.rs
@@ -14,6 +14,7 @@
 
 mod convert;
 mod distance;
+pub(crate) mod impl_conv;
 
 use std::sync::Arc;
